==> 2. Unsupervised Sampling.ipynb <==
import os
os.environ["KMP_DUPLICATE_LIB_OK"]="TRUE"

# =========================================================================================
# Libraries
# =========================================================================================
import gc
import time
import math
import random
import warnings
warnings.filterwarnings("ignore")
import numpy as np
import pandas as pd
from tqdm.auto import tqdm

# Custom libraries
from utils.unsupervised_utils import read_data, get_neighbors, build_training_set
from utils.utils import read_config
from utils.metrics import get_pos_score, get_f2_score

%env TOKENIZERS_PARALLELISM=false
%env TRANSFORMERS_NO_ADVISORY_WARNINGS=true

config = read_config()
DATA_PATH = "../raw_data/"

for mode in [
    "test",
    "train"
]:
    
    print(f"FOR MODE - {mode}")
    
    # Read data
    topics, content, correlations, _ = read_data(data_path=DATA_PATH,
                                              config_obj=config,
                                              read_mode=mode)
    
    # Run nearest neighbors
    topics, content = get_neighbors(topic_df=topics,
                                    content_df=content,
                                    config_obj=config)
    
    # Merge with target and comput max positive score
    topics = topics.merge(correlations, how = 'inner', left_on = ['id'], right_on = ['topic_id'])
    pos_score = get_pos_score(topics['content_ids'], topics['predictions'])
    f2_score = get_f2_score(topics['content_ids'], topics['predictions'])
    print(f'Max Recall @{config["unsupervised_model"]["top_n"]}: {pos_score}')
    print(f'F2-Score @{config["unsupervised_model"]["top_n"]}: {f2_score}')
    
    # Set id as index for content
    content.set_index('id', inplace = True)
    
    # Build training set
    df = build_training_set(topic_df=topics,
                            content_df=content)
    print(f'Our {mode} set has {len(df)} rows')
    
    # Save train set to disk to train on another notebook
    df.to_parquet(f'./generated_files/unsupervised_{mode}.parquet')
    df.head()
    
    del df, topics, content, correlations
    gc.collect()
